After testing Return proper error codes on listen failure
don't close serverFd (closed stdout)

use size_t for read/write sizes

+++ bsock.h
@@ -1,4 +1,5 @@
 #include <stdint.h>
+#include <stdlib.h>
 
 #define CONNECT_SUCCESS 1
 #define CONNECT_FAIL 	0
@@ -8,6 +9,6 @@
 
 int bsock_listenTCP(int portno);
 int bsock_connectTCP(const char * ip, int portno);
-int bsock_readTCP( uint8_t *bytes, uint32_t size);
-int bsock_writeTCP( uint8_t *bytes, uint32_t size);
+size_t bsock_readTCP( uint8_t *bytes, uint32_t size);
+size_t bsock_writeTCP( uint8_t *bytes, uint32_t size);
 void bsock_closeTCP();
